requirements.txt

aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bitarray==3.4.2
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
colorama==0.4.6
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.6.2
hexbytes==1.3.1
idna==3.10
multidict==6.4.4
parsimonious==0.10.0
pillow==11.2.1
propcache==0.3.1
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
pyunormalize==16.0.0
pywin32==310
qrcode==8.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
toolz==1.0.0
types-requests==2.32.0.20250602
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
web3==7.12.0
websockets==15.0.1
yarl==1.20.0
